requirements.txt
disnake==2.7.0
python-dotenv==0.21.0
mysql-connector-python~=8.0.31